Revert "Aggregate data in series as float32."
This reverts commit a1168bdffe33d07960fb533fc3ab44cde3249463.

+++ src/datasets/utils.py
@@ -518,7 +518,7 @@ def load_from_tsfile_to_dataframe(full_file_path_and_name, return_separate_X_and
                             if dimension:
                                 data_series = dimension.split(",")
                                 data_series = [float(i) for i in data_series]
-                                instance_list[dim].append(pd.Series(data_series, dtype=np.float16))
+                                instance_list[dim].append(pd.Series(data_series))
                             else:
                                 instance_list[dim].append(pd.Series())
